+++ b4n/ui/views/resources/view.rs
@@ -528,7 +528,7 @@ impl ResourcesView {
             highlighted,
             namespace,
             resource,
-            filter: self.table.list.table.get_filter().map(String::from),
+            filter: self.table.list.table.filter().map(String::from),
             sort_info: self.table.list.table.table.header.sort_info(),
             offset: self.table.list.table.offset(),
         });